requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
attrs
certifi
charset-normalizer
Deprecated
discord.py
et-xmlfile
frozenlist
idna
img2pdf
lxml
multidict
openpyxl
packaging
pikepdf
pillow
plyer
pydantic
pydantic_core
python-dotenv
PyPDF2
requests
typing_extensions
urllib3
webcolors
wrapt
yarl
